move logic into utils function, utils function still needs refactoring
